Split out backend.go
